server/pyproject.toml
[tool.poetry]
name = "kronos"
version = "0.1.0"
description = "Dependecy set for the Kronos Server platform."
authors = ["Emanuele Perrella <emanuele@devais.it>"]


[[tool.poetry.source]]
name = "gitlab"
url = "https://gitlab.com/api/v4/projects/29417850/packages/pypi/simple"

[tool.poetry.dependencies]
python = "^3.7"
uvicorn = "^0.13.4"
fastapi = "^0.63.0"
python-multipart = "^0.0.5"
requests = "^2.23.0"
tenacity = "^7.0.0"
pydantic = "^1.4"
gunicorn = "^20.0.4"
psycopg2-binary = "^2.8.5"
alembic = "^1.4.2"
sqlalchemy = "^1.4.7"
pytest = "^6.2.3"
pymysql = "^1.0.2"
paho-mqtt = "^1.5.1"
sqlalchemy-utils = "^0.37.0"
networkx = "^2.5"
matplotlib = "^3.3.4"
aiocache = "^0.11.1"
aioredis = "^1.3.1"
starlette-exporter = "^0.7.0"
pyaml = "^20.4.0"
pluggy = "^1.0.0"
python-jose = { extras = ["cryptography"], version = "^3.1.0" }
faust-utils = "^0.3"
py-healthcheck = "^1.10.1"
confluent-kafka = "^1.7.0"
httpcore = "<=0.15"

[tool.poetry.dev-dependencies]
mypy = "^0.812"
black = "^20.8b1"
isort = "^5.8.0"
autoflake = "^1.3.1"
flake8 = "^3.7.9"
pytest = "^6.2.3"
sqlalchemy-stubs = "^0.4"
pytest-cov = "^2.8.1"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

